chore: move metrics tracker to shared metrics utilities
Relocate createMetricsTracker from src/plugin/src/reporting to src/shared/metrics so instrumentation utilities are reusable beyond the plugin. Update imports, barrels, and tests to point at the shared entry point.
