+++ src/shared/metrics.js
@@ -1,0 +1,1 @@
+export * from "./metrics/index.js";
