+++ src/shared/metrics/metrics-tracker.js
@@ -1,8 +1,5 @@
-import {
-    getNonEmptyString,
-    normalizeStringList
-} from "../../../shared/string-utils.js";
-import { getOrCreateMapEntry } from "../../../shared/object-utils.js";
+import { getNonEmptyString, normalizeStringList } from "../string-utils.js";
+import { getOrCreateMapEntry } from "../object-utils.js";
 
 const hasHrtime = typeof process?.hrtime?.bigint === "function";
